+++ CHANGELOG.md
@@ -4,6 +4,21 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [v0.0.2] - Unreleased
+### Added
+- Windows traceroute module
+- Windows Ping Sweep, TCP and UDP nmap modules
+- Windows Defender module
+- Windows Metasploit module
+- Windows Gophish module
+- Linux dirb module 
+- Linux Covenant module
+- Linux Gophish module
+
+### Fixed
+- Linux shared folder permissions
+
+
 ## [v0.0.1] - 2021-07-16
 ### Added
 - New module execution verb "pivot" used to execute modules that require a redirection to an external application
@@ -27,6 +42,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Distribution-Server credentials generated at first deploy
 - Herd-View default user removed
 - npm dependencies updated in different package-lock.json files
+
 
 ## [v0.0.1c] - Unreleased
 ### Added